Move verify_subtype_sizes into its own file
