+++ drake/systems/robotInterfaces/verify_subtype_sizes.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "drake/lcmt_qp_controller_input.hpp"
+
+void verifySubtypeSizes(
+    // NOLINTNEXTLINE(runtime/references)
+    drake::lcmt_support_data& support_data);
+void verifySubtypeSizes(
+    // NOLINTNEXTLINE(runtime/references)
+    drake::lcmt_qp_controller_input& qp_input);
