+++ drake/util/BUILD.bazel
@@ -31,7 +31,6 @@ drake_cc_library(
         "//drake/lcmtypes:piecewise_polynomial",
         "//drake/lcmtypes:polynomial",
         "//drake/lcmtypes:polynomial_matrix",
-        "//drake/lcmtypes:qp_controller_input",
         "//drake/math:geometric_transform",
         "@eigen",
         "@lcmtypes_bot2_core",
